+++ graph/index.js
@@ -1,4 +1,5 @@
 const getNodeAlias = require('./get_node_alias');
 const getScoredNodes = require('./get_scored_nodes');
+const getSeedNodes = require('./get_seed_nodes');
 
-module.exports = {getNodeAlias, getScoredNodes};
+module.exports = {getNodeAlias, getScoredNodes, getSeedNodes};