+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 3.7.0
+
+- `getLiquidity`: Remove support for `min_node_score` due to API EOL
+- `getSeedNodes`: Add method to get some recommended starting nodes
+
 ## 3.6.1
 
 - `updateFeeRate`: Fix issue where fee rate would not update to zero